+++ src/videoframe.hpp
@@ -1,7 +1,7 @@
 /*
  * This file is part of Bino, a 3D video player.
  *
- * Copyright (C) 2022, 2023, 2024
+ * Copyright (C) 2022, 2023, 2024, 2025
  * Martin Lambers <marlam@marlam.de>
  *
  * This program is free software; you can redistribute it and/or modify
@@ -50,12 +50,19 @@ public:
         Storage_Image   // QImage
     };
 
-    enum YUVSpace {
+    enum ColorSpace {
         // see shader-color.frag.glsl
-        YUV_BT601 = 1,
-        YUV_BT709 = 2,
-        YUV_AdobeRgb = 3,
-        YUV_BT2020 = 4
+        CS_BT601 = 1,
+        CS_BT709 = 2,
+        CS_AdobeRgb = 3,
+        CS_BT2020 = 4
+    };
+
+    enum ColorTransfer {
+        // see shader-color.frag.glsl
+        CT_NOOP = 1,
+        CT_ST2084 = 2,
+        CT_STD_B67 = 3
     };
 
     /* This is a shallow copy of the original QVideoFrame: */
@@ -73,8 +80,10 @@ public:
     enum Storage storage;
     // for mapped and copied data:
     QVideoFrameFormat::PixelFormat pixelFormat;
-    bool yuvValueRangeSmall;
-    enum YUVSpace yuvSpace;
+    bool colorRangeSmall;
+    enum ColorSpace colorSpace;
+    enum ColorTransfer colorTransfer;
+    float masteringWhite;
     int planeCount; // 1-3
     int bytesPerLine[3];
     int bytesPerPlane[3];