+++ NEWS.md
@@ -1,3 +1,11 @@
+Version 2.5:
+- Improved meta data detection via ffprobe (if available).
+
+Version 2.4:
+- Added support for HDR video input.
+- Added support for zooming in surround videos.
+- Increased minimum Qt version to 6.6.
+
 Version 2.3:
 - Added support for slideshow mode.
 - Improved support for image file formats, including MPO.